dhcp4: implement basic client state machine
Implement the basic DHCP4 client state machine, including all the
transitions as defined by RFC2131. The underlying calls are no-ops so far
and no real data is passed on. Only the state detection and transition is
implemented.

Furthermore, this adds the epoll+timerfd integration to make the API easy
to use and never expose non-constant file-descriptors.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>